fix: Added lightweight App interface
